Fix all remaining compilation warnings - Add (void)parameter; to suppress unused parameter warnings in main.cpp and ftp_server_config.cpp - Fix sign comparison warnings in logger.cpp with proper type casting - Add missing getCurrentTimestamp method declaration and implementation - Clean build with zero warnings and errors

+++ src/utils/logger.cpp
@@ -320,14 +320,14 @@ void Logger::updatePerformanceMetrics(const std::chrono::steady_clock::time_poin
     
     // Handle atomic min/max operations properly
     uint64_t current_max = max_log_time_.load();
-    while (duration_us > current_max && 
-           !max_log_time_.compare_exchange_weak(current_max, duration_us)) {
+    while (static_cast<uint64_t>(duration_us) > current_max && 
+           !max_log_time_.compare_exchange_weak(current_max, static_cast<uint64_t>(duration_us))) {
         // Loop until we successfully update the max value
     }
     
     uint64_t current_min = min_log_time_.load();
-    while (duration_us < current_min && 
-           !min_log_time_.compare_exchange_weak(current_min, duration_us)) {
+    while (static_cast<uint64_t>(duration_us) < current_min && 
+           !min_log_time_.compare_exchange_weak(current_min, static_cast<uint64_t>(duration_us))) {
         // Loop until we successfully update the min value
     }
